more permissive conversion of blockquotes to divs
previously we forced blockquotes to begin with headers that contained a
single class that was in a list of allowed classes, if they were to be
converted to a div.

now any blockquote will be converted to a div provided it begins with a
header that has attributes containing at least one class.