doc/template: Move lesson-template docs to a subdirectory
We don't want them cluttering lesson repositories or creating merge
conflicts (e.g. if the README.md is adjusted in both core and the
downstream lesson).  However, we do want the docs in the core branch
(so we can land changes along with their docs).  Shifting these to a
subdirectory gives us a way to keep the docs in core while still
making them easy to ignore (or remove) if downstream lesson authors
aren't interested in them.