Use unified CSS files from site
This goes with jpallen/site#c40071f, and is a proof of concept that lesson can
share the same CSS as site. I haven't looked at merging slideshow.css yet
though.  This commit brings the CSS here up to date with the at site (if
c40071f is merged) although doesn't address how we'll keep these in sync in
future.