.update-copyright.conf: Enable per-file copyright blurbs
I expect the long blurbs will only go in "programs and other
software", so I've just used the MIT license for the content.  We'll
need fancier logic if we want MIT blurbs in the code and CC BY 4.0
blurbs in the instructional content.

I also added ignored entries to avoid bumping copyright information on
non-text files or files we've copy-pasted from other projects.