Add client-side hooks for lesson maintainers to regenerate HTML and update zip files.
This enhancement grew out of
[issue
94](https://github.com/swcarpentry/python-novice-inflammation/issues/94)
in the python-novice-inflammation repository and the discussion in a
[shell-novice PR](https://github.com/swcarpentry/shell-novice/pull/141).
It creates a maintainer_hooks directory in the tools directory. Lesson
maintainers can use these hooks by symlinking to them in the .git/hooks
directory. The pre-commit hook sets a "flag" file called .commit and the
post-commit hook detects the presence of this flag and regenerates HTML
and updates the data zip file if necessary. It then amends the previous
commit by adding these files to the commit and keeping the same message.